fix job seeder

+++ database/migrations/2024_04_21_043637_create_jobs_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->integer('count');
             $table->integer('Age');
             $table->text('gender');
-            $table->float('salary');
+            $table->integer('salary');
            $table->unsignedBigInteger('section_company_id');
            $table->foreign('section_company_id')->references('id')->on('section_companies')->onDelete('cascade');
             $table->timestamps();
